# packages/api-contract/openapi.yaml
openapi: 3.1.0
info:
  title: AI Operational POS API
  version: 0.1.0
  license:
    name: Proprietary
    identifier: LicenseRef-UNLICENSED
  description: Phase 0 foundation contract with the backend voice transcription slice. Product, auth, sales, inventory, and intent handlers are not declared until implemented.
servers:
  - url: http://127.0.0.1:8080
security: []
paths:
...
  /api/v1/customers:
    get:
      operationId: listCustomers
      summary: List active customers for the tenant
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
          description: Filter by name, phone, or email (case-insensitive).
      responses:
        '200':
          description: Tenant customer list ordered by name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessCustomerList'
        '404':
          description: Route not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      operationId: createCustomer
      summary: Register a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertCustomerRequest'
      responses:
        '201':
          description: Customer created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessCustomer'
        '400':
          description: Customer name missing or invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/customers/{id}:
    get:
      operationId: getCustomerById
      summary: Get a customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Customer detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessCustomer'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      operationId: updateCustomer
      summary: Update a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertCustomerRequest'
      responses:
        '200':
          description: Customer updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessCustomer'
        '400':
          description: Payload invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Customer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/pets:
    get:
      operationId: listPets
      summary: List active pets for the tenant
      parameters:
        - name: customer_id
          in: query
          required: false
          schema:
            type: string
          description: Filter by owner customer ID.
      responses:
        '200':
          description: Tenant pet list ordered by name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessPetList'
        '404':
          description: Route not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      operationId: createPet
      summary: Register a new pet for a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertPetRequest'
      responses:
        '201':
          description: Pet created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessPet'
        '422':
          description: Payload invalid or owner not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/pets/{id}:
    get:
      operationId: getPetById
      summary: Get a pet by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Pet detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessPet'
        '404':
          description: Pet not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      operationId: updatePet
      summary: Update a pet
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertPetRequest'
      responses:
        '200':
          description: Pet updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessPet'
        '404':
          description: Pet not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Payload invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error, request_id]
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
        request_id:
          type: string
    ErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          description: Optional machine-readable context for the error.
...
    Customer:
      type: object
      required: [id, tenant_id, name, is_active]
      properties:
        id:
          type: string
        tenant_id:
          type: string
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        notes:
          type: string
        is_active:
          type: boolean
    UpsertCustomerRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        address:
          type: string
        notes:
          type: string
    SuccessCustomerList:
      type: object
      required: [data, request_id]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Customer'
        request_id:
          type: string
    SuccessCustomer:
      type: object
      required: [data, request_id]
      properties:
        data:
          $ref: '#/components/schemas/Customer'
        request_id:
          type: string
    Pet:
      type: object
      required: [id, tenant_id, customer_id, name, is_active]
      properties:
        id:
          type: string
        tenant_id:
          type: string
        customer_id:
          type: string
        customer_name:
          type: string
        name:
          type: string
        species:
          type: string
        breed:
          type: string
        birth_date:
          type: string
          format: date
        gender:
          type: string
        weight_kg:
          type: number
        color:
          type: string
        allergies:
          type: string
        notes:
          type: string
        is_active:
          type: boolean
    UpsertPetRequest:
      type: object
      required: [customer_id, name]
      properties:
        customer_id:
          type: string
        name:
          type: string
        species:
          type: string
        breed:
          type: string
        birth_date:
          type: string
          format: date
        gender:
          type: string
        weight_kg:
          type: number
        color:
          type: string
        allergies:
          type: string
        notes:
          type: string
    SuccessPetList:
      type: object
      required: [data, request_id]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Pet'
        request_id:
          type: string
    SuccessPet:
      type: object
      required: [data, request_id]
      properties:
        data:
          $ref: '#/components/schemas/Pet'
        request_id:
          type: string
...